@@ -27,9 +27,9 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
-  flutter_staggered_grid_view: "^0.3.2"
-  intl: ^0.16.1
-  font_awesome_flutter: ^8.11.0
+  flutter_staggered_grid_view: ^0.4.0
+  intl: ^0.17.0
+  font_awesome_flutter: ^9.0.0
   shared_preferences: any
 
 dev_dependencies:
