@@ -15,14 +15,16 @@ class MusicasWebClient {
   }
 
   Future<List<Artista>> retornaArtistas() async {
-    final Response response = await client.get(await sharedPref.readString('ip') + 'listaArtista').timeout(Duration(seconds: 5));
+    final uri = new Uri.http(await sharedPref.readString('ip'), '/listaArtista');
+    final Response response = await client.get(uri).timeout(Duration(seconds: 5));
     final List<dynamic> decodedJson = jsonDecode(response.body);
 
     return decodedJson.map((dynamic json) => Artista.fromJson(json)).toList();
   }
 
   Future<List<Musica>> retornaMusicas(String artista) async {
-    final Response response = await client.get(await sharedPref.readString('ip') + 'listaMusica' + '?Artista=' + artista).timeout(Duration(seconds: 5));
+    final uri = new Uri.http(await sharedPref.readString('ip'), '/listaMusica', {"Artista": artista});
+    final Response response = await client.get(uri).timeout(Duration(seconds: 5));
     final List<dynamic> decodedJson = jsonDecode(response.body);
 
     return decodedJson.map((dynamic json) => Musica.fromJson(json)).toList();
@@ -30,8 +32,8 @@ class MusicasWebClient {
 
   Future<Success> reproduzirMusicas(String artista, String musica) async {
     final String selecionaMusica = jsonEncode({'Artista': artista, 'Musica': musica});
-
-    final Response response = await client.post(await sharedPref.readString('ip') + 'selecionaMusica',
+    final uri = new Uri.http(await sharedPref.readString('ip'), '/selecionaMusica');
+    final Response response = await client.post(uri,
         headers: {'Content-type' : 'application/json'},
         body: selecionaMusica
     );
