@@ -14,13 +14,15 @@ class ParametrosWebClient {
   }
 
   Future<Parametros> retornaParametros() async {
-    final Response response = await client.get(await sharedPref.readString('ip') + 'getParametro').timeout(Duration(seconds: 5));
+    final uri = new Uri.http(await sharedPref.readString('ip'), '/getParametro');
+    final Response response = await client.get(uri).timeout(Duration(seconds: 5));
 
     return Parametros.fromJson(jsonDecode(response.body));
   }
 
   Future<Success> salvaParametros(Parametros parametros) async {
-    final Response response = await client.post(await sharedPref.readString('ip') + 'setParametro',
+    final uri = new Uri.http(await sharedPref.readString('ip'), '/setParametro');
+    final Response response = await client.post(uri,
         headers: {'Content-type' : 'application/json'},
         body: jsonEncode(parametros.toJson()),
     ).timeout(Duration(seconds: 5));
